[b/381400224] Rename ClouderaMetadata connector to HadoopMetadata
